+++ packages/terra-core-docs/src/terra-dev-site/doc/form-input/Input.1.doc.mdx
@@ -1,19 +1,19 @@
 import { Badge } from 'terra-form-input/package.json?dev-site-package';
 
-import BlankExample from '../common/BlankExample?dev-site-example';
-import NumberInputExample from '../common/NumberInputExample?dev-site-example';
-import ControlledDefaultExample from '../example/controlled/DefaultExample?dev-site-example';
-import ControlledDisabledExample from '../example/controlled/DisabledExample?dev-site-example';
-import ControlledDefaultInvalidExample from '../example/controlled/DefaultInvalidExample?dev-site-example';
-import ControlledDefaultIncompleteExample from '../example/controlled/DefaultIncompleteExample?dev-site-example';
-import ControlledDisabledInvalidExample from '../example/controlled/DisabledInvalidExample?dev-site-example';
-import InputProgrammaticSet from '../example/controlled/InputProgrammaticSet?dev-site-example';
-import UncontrolledDefaultExample from '../example/uncontrolled/DefaultExample?dev-site-example';
-import UncontrolledDisabledExample from '../example/uncontrolled/DisabledExample?dev-site-example';
-import UncontrolledDefaultInvalidExample from '../example/uncontrolled/DefaultInvalidExample?dev-site-example';
-import UncontrolledDefaultIncompleteExample from '../example/uncontrolled/DefaultIncompleteExample?dev-site-example';
-import UncontrolledInvalidWithIncompleteExample from '../example/uncontrolled/InvalidWithIncompleteExample?dev-site-example';
-import UncontrolledDisabledInvalidExample from '../example/uncontrolled/DisabledInvalidExample?dev-site-example';
+import BlankExample from './common/BlankExample?dev-site-example';
+import NumberInputExample from './common/NumberInputExample?dev-site-example';
+import ControlledDefaultExample from './example/controlled/DefaultExample?dev-site-example';
+import ControlledDisabledExample from './example/controlled/DisabledExample?dev-site-example';
+import ControlledDefaultInvalidExample from './example/controlled/DefaultInvalidExample?dev-site-example';
+import ControlledDefaultIncompleteExample from './example/controlled/DefaultIncompleteExample?dev-site-example';
+import ControlledDisabledInvalidExample from './example/controlled/DisabledInvalidExample?dev-site-example';
+import InputProgrammaticSet from './example/controlled/InputProgrammaticSet?dev-site-example';
+import UncontrolledDefaultExample from './example/uncontrolled/DefaultExample?dev-site-example';
+import UncontrolledDisabledExample from './example/uncontrolled/DisabledExample?dev-site-example';
+import UncontrolledDefaultInvalidExample from './example/uncontrolled/DefaultInvalidExample?dev-site-example';
+import UncontrolledDefaultIncompleteExample from './example/uncontrolled/DefaultIncompleteExample?dev-site-example';
+import UncontrolledInvalidWithIncompleteExample from './example/uncontrolled/InvalidWithIncompleteExample?dev-site-example';
+import UncontrolledDisabledInvalidExample from './example/uncontrolled/DisabledInvalidExample?dev-site-example';
 
 import FormInputPropsTable from 'terra-form-input/src/Input?dev-site-props-table';
