+++ packages/terra-core-docs/src/terra-dev-site/doc/form-input/InputField.2.doc.mdx
@@ -1,9 +1,9 @@
 import { Badge } from 'terra-form-input/package.json?dev-site-package';
 
-import InputFieldExample from '../example/InputField?dev-site-example';
-import RequiredInputFieldExample from '../example/RequiredInputField?dev-site-example';
-import IncompleteInputFieldExample from '../example/IncompleteInputField?dev-site-example';
-import NumberInputFieldExample from '../example/NumberInputField?dev-site-example';
+import InputFieldExample from './example/InputField?dev-site-example';
+import RequiredInputFieldExample from './example/RequiredInputField?dev-site-example';
+import IncompleteInputFieldExample from './example/IncompleteInputField?dev-site-example';
+import NumberInputFieldExample from './example/NumberInputField?dev-site-example';
 
 import FormInputFieldPropsTable from 'terra-form-input/src/InputField?dev-site-props-table';
