+++ packages/terra-core-docs/src/terra-dev-site/doc/form-input/example/uncontrolled/DefaultExample.jsx
