+++ packages/terra-core-docs/src/terra-dev-site/test/form-input/DisabledValueInput.test.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Input from '../../../Input';
+import Input from 'terra-form-input';
 import styles from './common/Input.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const IncompleteInput = () => (
+const DisabledValueInput = () => (
   <div className={cx('content-wrapper')}>
-    <Input required isIncomplete id="form-input-incomplete" ariaLabel="label" />
+    <Input disabled id="form-input-disabled-value" value="value" ariaLabel="label" />
   </div>
 );
 
-export default IncompleteInput;
+export default DisabledValueInput;
