+++ packages/terra-core-docs/src/terra-dev-site/test/form-input/IncompleteFormInput.test.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Input from '../../../Input';
+import Input from 'terra-form-input';
 import styles from './common/Input.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const DefaultInput = () => (
+const IncompleteInput = () => (
   <div className={cx('content-wrapper')}>
-    <Input id="form-input-default" ariaLabel="label" />
+    <Input required isIncomplete id="form-input-incomplete" ariaLabel="label" />
   </div>
 );
 
-export default DefaultInput;
+export default IncompleteInput;
