+++ packages/terra-core-docs/src/terra-dev-site/test/form-input/UnsupportedTypesInput.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Input from '../../../Input';
+import Input from 'terra-form-input';
 import styles from './common/Input.test.module.scss';
 
 const cx = classNames.bind(styles);
